+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The **need for configuration updates** is **marked bold**.
 ### Fixes
 
 - fixed invalid framwork policy in the bruno DTR setup ([#967](https://github.com/eclipse-tractusx/puris/pull/967))
+- Fixed Modal dialogs not updating when deleting entries for materials ([#984](https://github.com/eclipse-tractusx/puris/pull/984))
 
 ## v3.4.0